requirements.txt
# Core PySide6 Framework
PySide6>=6.4.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Data Analysis
pandas>=2.0.0
matplotlib>=3.7.0

# Scientific Computing
scipy>=1.11.0

# Material Design Theme
qt-material>=2.14

# Version Management
packaging>=23.0

# YAML Support
PyYAML>=6.0
